+++ packages/svelte/src/lib/components/checkbox/split-checkbox-group-props.svelte.ts
@@ -4,4 +4,13 @@ import type { UseCheckboxGroupProps } from './use-checkbox-group.svelte'
 const splitFn = createSplitProps<UseCheckboxGroupProps>()
 
 export const splitCheckboxGroupProps = <T extends UseCheckboxGroupProps>(props: T) =>
-  splitFn(props, ['defaultValue', 'value', 'onValueChange', 'disabled', 'invalid', 'readOnly', 'name'])
+  splitFn(props, [
+    'defaultValue',
+    'value',
+    'onValueChange',
+    'disabled',
+    'invalid',
+    'readOnly',
+    'name',
+    'maxSelectedValues',
+  ])