+++ packages/vue/src/components/checkbox/checkbox-group.types.ts
@@ -24,6 +24,10 @@ export interface GroupProps {
    * If `true`, the checkbox group is invalid
    */
   invalid?: boolean
+  /**
+   * The maximum number of selected values
+   */
+  maxSelectedValues?: number
 }
 
 export type GroupEmits = {